+++ src/NEventStore.Domain/Persistence/EventStore/EventStoreRepository.Async.cs
@@ -15,7 +15,7 @@ namespace NEventStore.Domain.Persistence.EventStore
             return aggregate;
         }
 
-        public async Task SaveAsync(string bucketId, IAggregate aggregate, Guid commitId, Action<IDictionary<string, object>>? updateHeaders, CancellationToken cancellationToken)
+        public async Task<ICommit?> SaveAsync(string bucketId, IAggregate aggregate, Guid commitId, Action<IDictionary<string, object>>? updateHeaders, CancellationToken cancellationToken)
         {
             Dictionary<string, object> headers = PrepareHeaders(aggregate, updateHeaders);
             while (true)
@@ -25,16 +25,16 @@ namespace NEventStore.Domain.Persistence.EventStore
 
                 try
                 {
-                    await stream.CommitChangesAsync(commitId, cancellationToken).ConfigureAwait(false);
+                    var commit = await stream.CommitChangesAsync(commitId, cancellationToken).ConfigureAwait(false);
                     aggregate.ClearUncommittedEvents();
-                    return;
+                    return commit;
                 }
                 catch (DuplicateCommitException)
                 {
                     stream.ClearChanges();
                     // Issue: #4 and test: when_an_aggregate_is_persisted_using_the_same_commitId_twice
                     // should we rethrow the exception here? or provide a feedback whether the save was successful ?
-                    return;
+                    return null;
                 }
                 catch (ConcurrencyException e)
                 {
@@ -67,9 +67,8 @@ namespace NEventStore.Domain.Persistence.EventStore
 
         private async Task<IEventStream> OpenStreamAsync(string bucketId, Guid id, int version, ISnapshot? snapshot, CancellationToken cancellationToken)
         {
-            IEventStream stream;
             var streamId = bucketId + "+" + id;
-            if (_streams.TryGetValue(streamId, out stream))
+            if (_streams.TryGetValue(streamId, out IEventStream stream))
             {
                 return stream;
             }