+++ src/NEventStore.Domain/Persistence/EventStore/EventStoreRepository.cs
@@ -40,7 +40,7 @@ namespace NEventStore.Domain.Persistence.EventStore
             return aggregate;
         }
 
-        public void Save(string bucketId, IAggregate aggregate, Guid commitId, Action<IDictionary<string, object>>? updateHeaders)
+        public ICommit? Save(string bucketId, IAggregate aggregate, Guid commitId, Action<IDictionary<string, object>>? updateHeaders)
         {
             Dictionary<string, object> headers = PrepareHeaders(aggregate, updateHeaders);
             while (true)
@@ -50,16 +50,16 @@ namespace NEventStore.Domain.Persistence.EventStore
 
                 try
                 {
-                    stream.CommitChanges(commitId);
+                    var commit = stream.CommitChanges(commitId);
                     aggregate.ClearUncommittedEvents();
-                    return;
+                    return commit;
                 }
                 catch (DuplicateCommitException)
                 {
                     stream.ClearChanges();
                     // Issue: #4 and test: when_an_aggregate_is_persisted_using_the_same_commitId_twice
                     // should we rethrow the exception here? or provide a feedback whether the save was successful ?
-                    return;
+                    return null;
                 }
                 catch (ConcurrencyException e)
                 {
@@ -127,9 +127,8 @@ namespace NEventStore.Domain.Persistence.EventStore
 
         private IEventStream OpenStream(string bucketId, Guid id, int version, ISnapshot? snapshot)
         {
-            IEventStream stream;
             var streamId = bucketId + "+" + id;
-            if (_streams.TryGetValue(streamId, out stream))
+            if (_streams.TryGetValue(streamId, out IEventStream stream))
             {
                 return stream;
             }