+++ Changelog.md
@@ -1,5 +1,10 @@
 # NEventStore.Domain Versions
 
+## 10.1.0
+
+- Updated NEventStore reference to version 10.1.0
+- `IRepository.Save` and `IRepository.SaveAsync` now return the persisted `ICommit`.
+
 ## 10.0.0
 
 - Async Methods [#17](https://github.com/NEventStore/NEventStore.Domain/issues/17)